feat: support multiple graph in one page
