+++ packages/core/src/components/Wrapper.tsx
@@ -1,4 +1,4 @@
-import { FC, PropsWithChildren } from 'react';
+import type { FC, PropsWithChildren } from 'react';
 import { useGraphInstance } from '../hooks/useGraphInstance';
 
 const Wrapper: FC<PropsWithChildren> = ({ children }) => {
